+++ engine/flink/tests/src/test/scala/pl/touk/nussknacker/extensions/SpelExtension.scala
@@ -1,0 +1,11 @@
+package pl.touk.nussknacker.extensions
+
+import pl.touk.nussknacker.engine.graph.expression.Expression
+
+trait SpelExtension {
+
+  implicit class SpelExpresion(expression: String) {
+    def spel(): Expression = Expression.spel(expression)
+  }
+
+}